docs: adds metadata for human and parcel indicators
Adds YAML files to define attributes and factors for human indicators and parcel characteristics.

Moves the RData file to an archive directory.

Updates the source path for esca_sql_queries.R in knb-lter-cap.653.qmd to use a local repository path instead of a relative path, and changes the argument in the get_parcel_characteristics() and get_human_indicators() functions from a string literal to a string literal.